pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mlx-benchmark"
version = "0.1.0"
description = "Benchmark LLMs on Apple MLX knowledge and coding tasks"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
dependencies = [
    "ollama>=0.4.0",
]

[project.optional-dependencies]
anthropic = ["anthropic>=0.40.0"]
openai = ["openai>=1.30.0"]
plot = ["matplotlib>=3.7.0"]
all = [
    "anthropic>=0.40.0",
    "openai>=1.30.0",
    "pyyaml>=6.0",
    "matplotlib>=3.7.0",
]
config = ["pyyaml>=6.0"]
dev = [
    "ruff>=0.4.0",
    "pytest>=8.0",
    "pyyaml>=6.0",
]

[project.scripts]
mlx-bench = "mlx_benchmark.cli:main"

[tool.setuptools.packages.find]
include = ["mlx_benchmark*"]

[tool.setuptools.package-data]
"mlx_benchmark.data" = ["*.jsonl"]

[tool.ruff]
line-length = 100
target-version = "py310"